Implement new process card
